python/request_inception_service/requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Django==4.2.1
django-environ==0.10.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
inflection==0.5.1
jsonschema==4.17.3
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
